Store all devices in a new Kind, and pass the list of all devices into JS. Next step: use that list to draw graphs for all devices.
